beginning to make protodep more useful/cleaner
